Remove producer mode Make redis+consumer and memory+standalone the only valid SDK modes

+++ spec/engine_spec.rb
@@ -10,7 +10,8 @@ describe SplitIoClient, type: :client do
       SplitIoClient::SplitFactory.new('',
                                       logger: Logger.new(log),
                                       cache_adapter: cache_adapter,
-                                      redis_namespace: 'test').client
+                                      redis_namespace: 'test',
+                                      mode: @mode).client
     end
 
     let(:log) { StringIO.new }
@@ -50,17 +51,18 @@ describe SplitIoClient, type: :client do
     end
 
     before :each do
-      Redis.new.flushall
+      Redis.new.flushall if @mode.equal?(:consumer)
     end
 
     after :each do
-      Redis.new.flushall
+      Redis.new.flushall if @mode.equal?(:consumer)
     end
+
+    before { @mode = cache_adapter.equal?(:memory) ? :standalone : :consumer }
 
     context '#get_treatment' do
       before do
-        stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
-          .to_return(status: 200, body: all_keys_matcher_json)
+        load_splits(all_keys_matcher_json)
       end
 
       it 'saves just one metric to Redis' do
@@ -145,8 +147,7 @@ describe SplitIoClient, type: :client do
 
     context '#get_treatments' do
       before do
-        stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
-          .to_return(status: 200, body: all_keys_matcher_json)
+        load_splits(all_keys_matcher_json)
       end
 
       it 'saves just one metric to Redis' do
@@ -182,38 +183,7 @@ describe SplitIoClient, type: :client do
 
     context 'all keys matcher' do
       before do
-        stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
-          .to_return(status: 200, body: all_keys_matcher_json)
-      end
-
-      context 'producer mode' do
-        subject do
-          SplitIoClient.configuration = nil
-          SplitIoClient::SplitFactory.new('',
-                                          logger: Logger.new('/dev/null'),
-                                          cache_adapter: cache_adapter,
-                                          redis_namespace: 'test',
-                                          mode: :producer).client
-        end
-
-        it 'stores splits' do
-          expect(subject.instance_variable_get(:@adapter).splits_repository.splits.size).to eq(1)
-        end
-      end
-
-      context 'consumer mode' do
-        subject do
-          SplitIoClient.configuration = nil
-          SplitIoClient::SplitFactory.new('',
-                                          logger: Logger.new('/dev/null'),
-                                          cache_adapter: cache_adapter,
-                                          redis_namespace: 'test',
-                                          mode: :consumer).client
-        end
-
-        it 'stores splits' do
-          expect(subject.instance_variable_get(:@adapter).splits_repository.splits.size).to eq(0)
-        end
+        load_splits(all_keys_matcher_json)
       end
 
       it 'validates the feature is on for all ids' do
@@ -229,13 +199,9 @@ describe SplitIoClient, type: :client do
 
     context 'in segment matcher' do
       before do
-        stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
-          .to_return(status: 200, body: segment_matcher_json)
-      end
+        load_segments(segments_json)
 
-      before do
-        stub_request(:get, 'https://sdk.split.io/api/segmentChanges/demo?since=-1')
-          .to_return(status: 200, body: segments_json)
+        load_splits(segment_matcher_json)
       end
 
       it 'validates the feature is on for all ids' do
@@ -309,13 +275,9 @@ describe SplitIoClient, type: :client do
 
     context 'get_treatments in segment matcher' do
       before do
-        stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
-          .to_return(status: 200, body: segment_matcher2_json)
-      end
+        load_segments(segments_json)
 
-      before do
-        stub_request(:get, 'https://sdk.split.io/api/segmentChanges/demo?since=-1')
-          .to_return(status: 200, body: segments_json)
+        load_splits(segment_matcher2_json)
       end
 
       it 'validates the feature is on for all ids' do
@@ -357,8 +319,7 @@ describe SplitIoClient, type: :client do
 
     context 'whitelist matcher' do
       before do
-        stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
-          .to_return(status: 200, body: whitelist_matcher_json)
+        load_splits(whitelist_matcher_json)
       end
 
       it 'validates the feature is on for all ids' do
@@ -372,8 +333,7 @@ describe SplitIoClient, type: :client do
 
     context 'dependency matcher' do
       before do
-        stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
-          .to_return(status: 200, body: dependency_matcher_json)
+        load_splits(dependency_matcher_json)
       end
 
       it 'returns on treatment' do
@@ -390,8 +350,7 @@ describe SplitIoClient, type: :client do
 
     context 'killed feature' do
       before do
-        stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
-          .to_return(status: 200, body: killed_json)
+        load_splits(killed_json)
       end
 
       it 'returns default treatment for killed splits' do
@@ -403,13 +362,9 @@ describe SplitIoClient, type: :client do
 
     context 'deleted segment' do
       before do
-        stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
-          .to_return(status: 200, body: segment_deleted_matcher_json)
-      end
+        load_segments(segments_json)
 
-      before do
-        stub_request(:get, 'https://sdk.split.io/api/segmentChanges/demo?since=-1')
-          .to_return(status: 200, body: segments_json)
+        load_splits(segment_deleted_matcher_json)
       end
 
       it 'returns control for deleted splits' do
@@ -459,8 +414,7 @@ describe SplitIoClient, type: :client do
       end
 
       before do
-        stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
-          .to_return(status: 200, body: impressions_test_json)
+        load_splits(impressions_test_json)
       end
 
       it 'returns correct impressions for get_treatments' do
@@ -485,11 +439,14 @@ describe SplitIoClient, type: :client do
       context 'when impressions are disabled' do
         subject do
           SplitIoClient.configuration = nil
+          mode = cache_adapter.equal?(:memory) ? :standalone : :consumer
+
           SplitIoClient::SplitFactory.new('',
                                           logger: Logger.new('/dev/null'),
                                           cache_adapter: cache_adapter,
                                           redis_namespace: 'test',
-                                          impressions_queue_size: -1).client
+                                          impressions_queue_size: -1,
+                                          mode: mode).client
         end
         let(:impressions) { subject.instance_variable_get(:@impressions_repository).batch }
 
@@ -511,8 +468,7 @@ describe SplitIoClient, type: :client do
 
       context 'traffic allocations' do
         before do
-          stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
-            .to_return(status: 200, body: traffic_allocation_json)
+          load_splits(traffic_allocation_json)
         end
 
         it 'returns expected treatment' do
@@ -544,8 +500,7 @@ describe SplitIoClient, type: :client do
 
     describe 'client destroy' do
       before do
-        stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
-          .to_return(status: 200, body: all_keys_matcher_json)
+        load_splits(all_keys_matcher_json)
       end
 
       it 'returns control' do
@@ -689,6 +644,54 @@ describe SplitIoClient, type: :client do
       end
     end
   end
+
+  context 'SDK modes' do
+    let(:all_keys_matcher_json) do
+      File.read(File.join(SplitIoClient.root, 'spec/test_data/splits/engine/all_keys_matcher.json'))
+    end
+
+    before do
+      stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
+        .to_return(status: 200, body: all_keys_matcher_json)
+    end
+
+    context 'standalone mode' do
+      subject do
+        SplitIoClient.configuration = nil
+        SplitIoClient::SplitFactory.new('',
+                                        logger: Logger.new('/dev/null'),
+                                        cache_adapter: :memory,
+                                        mode: :standalone).client
+      end
+
+      it 'stores splits' do
+        expect(subject.instance_variable_get(:@adapter).splits_repository.splits.size).to eq(1)
+      end
+    end
+
+    context 'consumer mode' do
+      before :each do
+        Redis.new.flushall
+      end
+
+      after :each do
+        Redis.new.flushall
+      end
+
+      subject do
+        SplitIoClient.configuration = nil
+        SplitIoClient::SplitFactory.new('',
+                                        logger: Logger.new('/dev/null'),
+                                        cache_adapter: :redis,
+                                        redis_namespace: 'test',
+                                        mode: :consumer).client
+      end
+
+      it 'does not store splits' do
+        expect(subject.instance_variable_get(:@adapter).splits_repository.splits.size).to eq(0)
+      end
+    end
+  end
 end
 
 describe SplitIoClient do
@@ -697,4 +700,40 @@ end
 
 describe SplitIoClient do
   include_examples 'engine specs', :redis
+end
+
+private
+
+def load_splits(splits_json)
+  if @mode.equal?(:standalone)
+    stub_request(:get, 'https://sdk.split.io/api/splitChanges?since=-1')
+      .to_return(status: 200, body: splits_json)
+  else
+    add_splits_to_repository(splits_json)
+  end
+end
+
+def load_segments(segments_json)
+  if @mode.equal?(:standalone)
+    stub_request(:get, 'https://sdk.split.io/api/segmentChanges/demo?since=-1')
+      .to_return(status: 200, body: segments_json)
+  else
+    add_segments_to_repository(segments_json)
+  end
+end
+
+def add_splits_to_repository(splits_json)
+  splits = JSON.parse(splits_json, symbolize_names: true)[:splits]
+
+  splits_repository = subject.instance_variable_get(:@splits_repository)
+
+  splits.each do |split|
+    splits_repository.add_split(split)
+  end
+end
+
+def add_segments_to_repository(segments_json)
+  segments_repository = subject.instance_variable_get(:@segments_repository)
+
+  segments_repository.add_to_segment(JSON.parse(segments_json, symbolize_names: true))
 end
